// libraries/common-ui/src/types.d.ts
export type Direction = "vertical" | "horizontal";
// ...
export type ChartValue = string | number | Date;

export type SeriesType = "bar" | "line" | "box-plot" | "pie" | "stack" | "area" | "agg-line";

export type ColorStandard = "series" | "axis" | "one";

export type AxisType =
  | "key"
  | "series"
  | "series-aggregation"
  | "series-aggregation-accumulate"
  | "series-aggregation-minmax";

export type AxisDataType = "string" | "number" | "Date";

export type TextType = "text" | "legend" | "label" | "total" | "html";

export type ShapeType = "circle" | "ractangle" | "series" | "square";

export type Aggregation =
  | "count"
  | "sum"
  | "min"
  | "max"
  | "avg"
  | "median"
  | "deviation"
  | "variance"
  | "quantileThree"
  | "quantileOne";

export type Location = "top" | "right" | "bottom" | "left" | "center";

export type Position = "start" | "middle" | "end";

// hanging = top, ideographic = bottom, middle = central
export type BaseLine = "hanging" | "ideographic" | "central";

export type Anchor = "start" | "middle" | "end";

export type CollisionTextType = "back" | "series";

export interface IChartData {
  [key: string]: any;
  key: ChartValue;
}

export interface INonNullMargin {
  [key: string]: any;
  top: number;
  right: number;
  bottom: number;
  left: number;
}

export interface IMargin {
  [key: string]: any;
  top?: number;
  right?: number;
  bottom?: number;
  left?: number;
}

export interface IChartValue {
  idx: number;
  keyData: ChartValue;
  valueData: number;
  valueField: string;
  outlier?: boolean;
  highlight?: boolean;
}

export interface IChartAggregation {
  valueData: number;
  valueField: string;
}

export interface IChartTooltipParams {
  type: "key" | "value";
  key: ChartValue;
  text: ChartValue;
  numberFormat?: (value: number, percent?: boolean) => string;
}

export interface IChartTextParams {
  value: ChartValue;
  numberFormat?: (value: number, percent?: boolean) => string;
}